third_party/WebKit/LayoutTests/fast/css/invalidation/explicit-inheritance-non-inherited.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<style>
html { background-color: green }
* { background-color: inherit }
</style>
<div id="root">
    <div>
        <span></span>
        <span></span>
        <span></span>
        <span></span>
    </div>
</div>
<script>
description("Inheritance-propagation should stop when there is no style diff.");

root.offsetTop; // force recalc

root.style.marginLeft = "1px";

if (window.internals)
    shouldBe("internals.updateStyleAndReturnAffectedElementCount()", "2");
</script>
